Bridge: split protocol and aes_gcm_siv bindings into separate files
Apart from helping avoid the top-level file growing infinitely, this
also allows using a dedicated Result type in each file instead of
writing SignalProtocolError or aes_gcm_siv::Error over and over.
